// front/src/backend_api/backend/models/CategoryResult.ts
export interface CategoryResult {
    /**
     * 
     * @type {string}
     * @memberof CategoryResult
     */
    categoryId: string;
    /**
     * 
     * @type {string}
     * @memberof CategoryResult
     */
    alternateName: string | null;
    /**
     * 
     * @type {string}
     * @memberof CategoryResult
     */
    color: string | null;
    /**
     * 
     * @type {number}
     * @memberof CategoryResult
     */
    minPoints: number;
    /**
     * 
     * @type {number}
     * @memberof CategoryResult
     */
    maxPoints: number;
    /**
     * 
     * @type {Date}
     * @memberof CategoryResult
     */
    startTime: Date;
    /**
     * 
     * @type {boolean}
     * @memberof CategoryResult
     */
    womenOnly: boolean;
    /**
     * 
     * @type {number}
     * @memberof CategoryResult
     */
    baseRegistrationFee: number;
    /**
     * 
     * @type {number}
     * @memberof CategoryResult
     */
    lateRegistrationFee: number;
    /**
     * 
     * @type {number}
     * @memberof CategoryResult
     */
    rewardFirst: number;
    /**
     * 
     * @type {number}
     * @memberof CategoryResult
     */
    rewardSecond: number;
    /**
     * 
     * @type {number}
     * @memberof CategoryResult
     */
    rewardSemi: number;
    /**
     * 
     * @type {number}
     * @memberof CategoryResult
     */
    rewardQuarter: number | null;
    /**
     * 
     * @type {number}
     * @memberof CategoryResult
     */
    maxPlayers: number;
    /**
     * 
     * @type {number}
     * @memberof CategoryResult
     */
    overbookingPercentage: number;
    /**
     * 
     * @type {number}
     * @memberof CategoryResult
     */
    entryCount: number;
    /**
     * 
     * @type {number}
     * @memberof CategoryResult
     */
    presentEntryCount: number;
    /**
     * 
     * @type {number}
     * @memberof CategoryResult
     */
    currentFee: number;
}